Move database thread into dedicated module.
