+++ src/database/database-thread.h
@@ -1,0 +1,15 @@
+/* Pi-hole: A black hole for Internet advertisements
+*  (c) 2019 Pi-hole, LLC (https://pi-hole.net)
+*  Network-wide ad blocking via your own hardware.
+*
+*  FTL Engine
+*  Database thread prototype
+*
+*  This file is copyright under the latest version of the EUPL.
+*  Please see LICENSE file for your rights under this license. */
+#ifndef DATABASE_THREAD_H
+#define DATABASE_THREAD_H
+
+void *DB_thread(void *val);
+
+#endif //DATABASE_THREAD_H
